Change Offset to use u32

+++ volpe-parser/src/offset.rs
@@ -2,18 +2,18 @@ use std::ops::{Add, AddAssign, Sub, SubAssign};
 
 #[derive(Clone, Copy, Default, PartialEq, PartialOrd, Eq, Ord, Debug)]
 pub struct Offset {
-    pub line: usize,
-    pub char: usize,
+    pub line: u32,
+    pub char: u32,
 }
 
 impl Offset {
-    pub fn new(line: usize, char: usize) -> Self {
+    pub fn new(line: u32, char: u32) -> Self {
         Self { line, char }
     }
     pub fn line() -> Self {
         Self { line: 1, char: 0 }
     }
-    pub fn char(char: usize) -> Self {
+    pub fn char(char: u32) -> Self {
         Self { line: 0, char }
     }
 }
